requirements.txt
Django==4.2.7
djangorestframework==3.14.0
sqlparse>=0.3.1
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
python-jose==3.3.0
cryptography==41.0.7
Pillow==10.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
djongo==1.3.6
pymongo==4.6.0
